Add parameter threshold ProfanityFilter and move utils under primitives.
.
