+++ RecDP/pyrecdp/primitives/operations/filter/token_num_filter.py
@@ -1,10 +1,10 @@
 import sys
 
 from pyrecdp.core.model_utils import get_model, prepare_model
-from pyrecdp.core.utils import get_words_from_document
 from pyrecdp.primitives.operations.base import LLMOPERATORS
 from pyrecdp.primitives.operations.filter.base import BaseFilter
 from pyrecdp.primitives.operations.filter.constant import HF_TOKENIZER
+from pyrecdp.primitives.operations.utils import get_words_from_document
 
 
 # This filter is referred from alibaba data juicer project
